fix tests for metadata writing
